Mixing Cauldrons are now mixed with spawners, various non-interactable knick knacks around the world are shuffled (like books and bottles). #3
Better method for getting the localURL (which also impacted
randomization, since it's used in the seed logic)

Make sure DXRando doesn't persist (despite it being transient), and make
sure the reference to it in Harry is cleared on PreTravel.

New Fixups module to make map changes.  For now, just making sure the
mixing cauldron in the Potions class stays where it is.